ai_interview_appback/ai_interview_app/requirements.txt
# requirements.txt - Project dependencies

fastapi>=0.111.0 # Web framework
uvicorn>=0.20.0 # ASGI server

pydantic>=2.0.0 # Data validation
pydantic-settings>=2.0.0 # Settings management

openai>=1.0.0 # For OpenAI API interaction

celery>=5.0.0 # Asynchronous task queue
redis>=4.0.0 # Redis client (used for Celery broker/backend)

python-multipart>=0.0.5 # Required for file uploads in FastAPI

# Document parsing libraries - choose based on expected file types
# For PDF:
PyMuPDF>=1.24.0
# For DOCX:
python-docx>=0.8.11

# Add other libraries as needed (e.g., database drivers, other services)
# asyncpg # If using PostgreSQL
# aiohttp # If using async HTTP client